DriveNTech/requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.5
defusedxml==0.7.1
Django==5.2.4
django-axes==8.0.0
django-cors-headers==4.7.0
django-filter==25.1
django-oauth-toolkit==3.0.1
django-ratelimit==4.1.0
django-redis==6.0.0
django-restframework==0.0.1
django-sslserver==0.22
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
httpie==3.2.4
idna==3.10
inflection==0.5.1
jwcrypto==1.5.6
kombu==5.5.4
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.6.3
oauthlib==3.3.1
packaging==25.0
pillow==11.3.0
prompt_toolkit==3.0.51
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.3.0
requests==2.32.4
requests-toolbelt==1.0.0
rich==14.1.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
